refactor: SearchDialog 필터 컴포넌트 통합 및 최적화
- 분리된 6개 필터 컴포넌트를 SearchDialog로 통합
- 각 필터 섹션을 useMemo로 메모이제이션하여 독립적인 리렌더링 보장
- 핸들러 함수들을 useCallback으로 메모이제이션하여 참조 안정화
- 개별 상태 관리 유지로 변경되는 영역만 리렌더링되도록 최적화
