--- src/compomnents/SearchDialog/MajorsFilter.tsx
@@ -1,64 +1,0 @@
-import { memo } from "react";
-import {
-  Box,
-  Checkbox,
-  CheckboxGroup,
-  FormControl,
-  FormLabel,
-  Stack,
-  Tag,
-  TagCloseButton,
-  TagLabel,
-  Wrap,
-} from "@chakra-ui/react";
-
-interface Props {
-  majors: string[];
-  allMajors: string[];
-  onChange: (values: string[]) => void;
-}
-
-const MajorsFilter = memo(({ majors, allMajors, onChange }: Props) => {
-  return (
-    <FormControl>
-      <FormLabel>전공</FormLabel>
-      <CheckboxGroup
-        colorScheme="green"
-        value={majors}
-        onChange={(values) => onChange(values as string[])}
-      >
-        <Wrap spacing={1} mb={2}>
-          {majors.map((major) => (
-            <Tag key={major} size="sm" variant="outline" colorScheme="blue">
-              <TagLabel>{major.split("<p>").pop()}</TagLabel>
-              <TagCloseButton
-                onClick={() => onChange(majors.filter((v) => v !== major))}
-              />
-            </Tag>
-          ))}
-        </Wrap>
-        <Stack
-          spacing={2}
-          overflowY="auto"
-          h="100px"
-          border="1px solid"
-          borderColor="gray.200"
-          borderRadius={5}
-          p={2}
-        >
-          {allMajors.map((major) => (
-            <Box key={major}>
-              <Checkbox key={major} size="sm" value={major}>
-                {major.replace(/<p>/gi, " ")}
-              </Checkbox>
-            </Box>
-          ))}
-        </Stack>
-      </CheckboxGroup>
-    </FormControl>
-  );
-});
-
-MajorsFilter.displayName = "MajorsFilter";
-
-export default MajorsFilter;
