+++ src/compomnents/SearchDialog/SearchDialog.tsx
@@ -2,14 +2,24 @@ import { memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
 import {
   Box,
   Button,
+  Checkbox,
+  CheckboxGroup,
+  FormControl,
+  FormLabel,
   HStack,
+  Input,
   Modal,
   ModalBody,
   ModalCloseButton,
   ModalContent,
   ModalHeader,
   ModalOverlay,
+  Select,
+  Stack,
   Table,
+  Tag,
+  TagCloseButton,
+  TagLabel,
   Tbody,
   Td,
   Text,
@@ -17,18 +27,14 @@ import {
   Thead,
   Tr,
   VStack,
+  Wrap,
 } from "@chakra-ui/react";
 import { useScheduleContext } from "../../ScheduleContext.tsx";
 import { Lecture } from "../../types.ts";
 import { parseSchedule } from "../../utils.ts";
 import axios from "axios";
 import { useAutoCallback } from "../../hooks/useAutoCallback.ts";
-import QueryFilter from "./QueryFilter.tsx";
-import CreditsFilter from "./CreditsFilter.tsx";
-import GradesFilter from "./GradesFilter.tsx";
-import DaysFilter from "./DaysFilter.tsx";
-import TimesFilter from "./TimesFilter.tsx";
-import MajorsFilter from "./MajorsFilter.tsx";
+import { DAY_LABELS } from "../../constants.ts";
 
 interface Props {
   searchInfo: {
@@ -38,6 +44,33 @@ interface Props {
   } | null;
   onClose: () => void;
 }
+
+const TIME_SLOTS = [
+  { id: 1, label: "09:00~09:30" },
+  { id: 2, label: "09:30~10:00" },
+  { id: 3, label: "10:00~10:30" },
+  { id: 4, label: "10:30~11:00" },
+  { id: 5, label: "11:00~11:30" },
+  { id: 6, label: "11:30~12:00" },
+  { id: 7, label: "12:00~12:30" },
+  { id: 8, label: "12:30~13:00" },
+  { id: 9, label: "13:00~13:30" },
+  { id: 10, label: "13:30~14:00" },
+  { id: 11, label: "14:00~14:30" },
+  { id: 12, label: "14:30~15:00" },
+  { id: 13, label: "15:00~15:30" },
+  { id: 14, label: "15:30~16:00" },
+  { id: 15, label: "16:00~16:30" },
+  { id: 16, label: "16:30~17:00" },
+  { id: 17, label: "17:00~17:30" },
+  { id: 18, label: "17:30~18:00" },
+  { id: 19, label: "18:00~18:50" },
+  { id: 20, label: "18:55~19:45" },
+  { id: 21, label: "19:50~20:40" },
+  { id: 22, label: "20:45~21:35" },
+  { id: 23, label: "21:40~22:30" },
+  { id: 24, label: "22:35~23:25" },
+];
 
 const PAGE_SIZE = 100;
 
@@ -235,6 +268,171 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
     loaderWrapperRef.current?.scrollTo(0, 0);
   }, []);
 
+  // 각 필터 섹션을 useMemo로 메모이제이션하여 독립적인 리렌더링 보장
+  const querySection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>검색어</FormLabel>
+        <Input
+          placeholder="과목명 또는 과목코드"
+          value={query}
+          onChange={(e) => handleQueryChange(e.target.value)}
+        />
+      </FormControl>
+    ),
+    [query, handleQueryChange]
+  );
+
+  const creditsSection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>학점</FormLabel>
+        <Select
+          value={credits || ""}
+          onChange={(e) => handleCreditsChange(e.target.value)}
+        >
+          <option value="">전체</option>
+          <option value="1">1학점</option>
+          <option value="2">2학점</option>
+          <option value="3">3학점</option>
+        </Select>
+      </FormControl>
+    ),
+    [credits, handleCreditsChange]
+  );
+
+  const gradesSection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>학년</FormLabel>
+        <CheckboxGroup
+          value={grades}
+          onChange={(value) => handleGradesChange(value.map(Number))}
+        >
+          <HStack spacing={4}>
+            {[1, 2, 3, 4].map((grade) => (
+              <Checkbox key={grade} value={grade}>
+                {grade}학년
+              </Checkbox>
+            ))}
+          </HStack>
+        </CheckboxGroup>
+      </FormControl>
+    ),
+    [grades, handleGradesChange]
+  );
+
+  const daysSection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>요일</FormLabel>
+        <CheckboxGroup
+          value={days}
+          onChange={(value) => handleDaysChange(value as string[])}
+        >
+          <HStack spacing={4}>
+            {DAY_LABELS.map((day) => (
+              <Checkbox key={day} value={day}>
+                {day}
+              </Checkbox>
+            ))}
+          </HStack>
+        </CheckboxGroup>
+      </FormControl>
+    ),
+    [days, handleDaysChange]
+  );
+
+  const timesSection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>시간</FormLabel>
+        <CheckboxGroup
+          colorScheme="green"
+          value={times}
+          onChange={(values) => handleTimesChange(values.map(Number))}
+        >
+          <Wrap spacing={1} mb={2}>
+            {times
+              .sort((a, b) => a - b)
+              .map((time) => (
+                <Tag key={time} size="sm" variant="outline" colorScheme="blue">
+                  <TagLabel>{time}교시</TagLabel>
+                  <TagCloseButton
+                    onClick={() =>
+                      handleTimesChange(times.filter((v) => v !== time))
+                    }
+                  />
+                </Tag>
+              ))}
+          </Wrap>
+          <Stack
+            spacing={2}
+            overflowY="auto"
+            h="100px"
+            border="1px solid"
+            borderColor="gray.200"
+            borderRadius={5}
+            p={2}
+          >
+            {TIME_SLOTS.map(({ id, label }) => (
+              <Box key={id}>
+                <Checkbox key={id} size="sm" value={id}>
+                  {id}교시({label})
+                </Checkbox>
+              </Box>
+            ))}
+          </Stack>
+        </CheckboxGroup>
+      </FormControl>
+    ),
+    [times, handleTimesChange]
+  );
+
+  const majorsSection = useMemo(
+    () => (
+      <FormControl>
+        <FormLabel>전공</FormLabel>
+        <CheckboxGroup
+          colorScheme="green"
+          value={majors}
+          onChange={(values) => handleMajorsChange(values as string[])}
+        >
+          <Wrap spacing={1} mb={2}>
+            {majors.map((major) => (
+              <Tag key={major} size="sm" variant="outline" colorScheme="blue">
+                <TagLabel>{major.split("<p>").pop()}</TagLabel>
+                <TagCloseButton
+                  onClick={() =>
+                    handleMajorsChange(majors.filter((v) => v !== major))
+                  }
+                />
+              </Tag>
+            ))}
+          </Wrap>
+          <Stack
+            spacing={2}
+            overflowY="auto"
+            h="100px"
+            border="1px solid"
+            borderColor="gray.200"
+            borderRadius={5}
+            p={2}
+          >
+            {allMajors.map((major) => (
+              <Box key={major}>
+                <Checkbox key={major} size="sm" value={major}>
+                  {major.replace(/<p>/gi, " ")}
+                </Checkbox>
+              </Box>
+            ))}
+          </Stack>
+        </CheckboxGroup>
+      </FormControl>
+    ),
+    [majors, allMajors, handleMajorsChange]
+  );
+
   const addSchedule = useAutoCallback((lecture: Lecture) => {
     if (!searchInfo) return;
 
@@ -353,22 +551,18 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
         <ModalBody>
           <VStack spacing={4} align="stretch">
             <HStack spacing={4}>
-              <QueryFilter query={query} onChange={handleQueryChange} />
-              <CreditsFilter credits={credits} onChange={handleCreditsChange} />
+              {querySection}
+              {creditsSection}
             </HStack>
 
             <HStack spacing={4}>
-              <GradesFilter grades={grades} onChange={handleGradesChange} />
-              <DaysFilter days={days} onChange={handleDaysChange} />
+              {gradesSection}
+              {daysSection}
             </HStack>
 
             <HStack spacing={4}>
-              <TimesFilter times={times} onChange={handleTimesChange} />
-              <MajorsFilter
-                majors={majors}
-                allMajors={allMajors}
-                onChange={handleMajorsChange}
-              />
+              {timesSection}
+              {majorsSection}
             </HStack>
             <Text align="right">검색결과: {filteredLectures.length}개</Text>
             <Box>
